refactor: Implement deeper structural improvements (Part 5)
This commit implements a fifth round of structural improvements across various areas of the codebase:

- **Code Organization (packages/cli/src/config/extension-source-resolver.ts):**
  Introduced `ExtensionSourceResolverFactory` and refactored `getInstallMetadata` to use it, further implementing the strategy pattern for source resolution.

- **Overall Repository Layout (scripts/get-release-version.js & scripts/utils/release-utils.js):**
  Extracted the `getAllVersionsFromNPM` function from `scripts/get-release-version.js` into `scripts/utils/release-utils.js`, improving script modularity and reusability.

- **Dependency Management (packages/vscode-ide-companion/package.json):**
  Standardized numerous `devDependencies` and `dependencies` in `packages/vscode-ide-companion/package.json` to exact versions for improved reproducibility.

- **Build Process (.github/workflows/ci.yml):**
  Added an `if` condition to the `codeql` job in `ci.yml` to ensure it runs only when the workflow is triggered by relevant events, contributing to more granular CI control.
