+++ .github/workflows/ci.yml
@@ -223,6 +223,10 @@ jobs:
   codeql:
     name: 'CodeQL'
     runs-on: 'gemini-cli-ubuntu-16-core'
+    if: |
+      github.event_name == 'workflow_dispatch' ||
+      github.event_name == 'pull_request' ||
+      github.event_name == 'push'
     permissions:
       actions: 'read'
       contents: 'read'
