+++ packages/vscode-ide-companion/package.json
@@ -113,23 +113,23 @@
     "validate:notices": "node ./scripts/validate-notices.js"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.19",
-    "@types/express": "^5.0.3",
+    "@types/cors": "2.8.19",
+    "@types/express": "5.0.3",
     "@types/node": "20.x",
-    "@types/vscode": "^1.99.0",
-    "@typescript-eslint/eslint-plugin": "^8.31.1",
-    "@typescript-eslint/parser": "^8.31.1",
-    "@vscode/vsce": "^3.6.0",
-    "esbuild": "^0.25.3",
-    "eslint": "^9.25.1",
-    "npm-run-all2": "^8.0.2",
-    "typescript": "^5.8.3",
-    "vitest": "^3.2.4"
+    "@types/vscode": "1.99.0",
+    "@typescript-eslint/eslint-plugin": "8.31.1",
+    "@typescript-eslint/parser": "8.31.1",
+    "@vscode/vsce": "3.6.0",
+    "esbuild": "0.25.3",
+    "eslint": "9.25.1",
+    "npm-run-all2": "8.0.2",
+    "typescript": "5.8.3",
+    "vitest": "3.2.4"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.15.1",
-    "cors": "^2.8.5",
-    "express": "^5.1.0",
-    "zod": "^3.25.76"
+    "@modelcontextprotocol/sdk": "1.15.1",
+    "cors": "2.8.5",
+    "express": "5.1.0",
+    "zod": "3.25.76"
   }
 }
